Remove duplicated code, create utils.microphysics
